fix image loading

+++ images/image_storagge_manager.h
@@ -1,11 +1,12 @@
 #ifndef IMAGE_STORAGGE_MANAGER_H
 #define IMAGE_STORAGGE_MANAGER_H
 
+#include <string>
 
 class ImageStorageManager {
 
   public:
-    void loadImage(const std::string& imagePath, float** imageData, int* width, int* height);
+    void loadImage(const std::string& imagePath, double** imageData, size_t* width, size_t* height);
 
 
 };
